Enable deleting athletes by -p flag (test cases not passed)
